+++ packages/block-library/src/breadcrumbs/block.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
-	"name": "core/post-breadcrumbs",
+	"name": "core/breadcrumbs",
 	"title": "Breadcrumbs",
 	"category": "theme",
 	"description": "Display a breadcrumb trail only for Pages, or for hierarchical post types. The block is useful to insert in the Pages template.",
@@ -60,5 +60,5 @@
 			"clientNavigation": true
 		}
 	},
-	"style": "wp-block-post-breadcrumbs"
+	"style": "wp-block-breadcrumbs"
 }
