+++ packages/block-library/src/breadcrumbs/style.scss
@@ -1,4 +1,4 @@
-.wp-block-post-breadcrumbs {
+.wp-block-breadcrumbs {
 	// This block has customizable padding, border-box makes that more predictable.
 	box-sizing: border-box;
