+++ packages/block-library/src/index.js
@@ -33,6 +33,7 @@ import * as accordionPanel from './accordion-panel';
 import * as archives from './archives';
 import * as avatar from './avatar';
 import * as audio from './audio';
+import * as breadcrumbs from './breadcrumbs';
 import * as button from './button';
 import * as buttons from './buttons';
 import * as calendar from './calendar';
@@ -87,7 +88,6 @@ import * as paragraph from './paragraph';
 import * as postAuthor from './post-author';
 import * as postAuthorName from './post-author-name';
 import * as postAuthorBiography from './post-author-biography';
-import * as postBreadcrumbs from './post-breadcrumbs';
 import * as postComment from './post-comment';
 import * as postCommentsCount from './post-comments-count';
 import * as postCommentsForm from './post-comments-form';
@@ -161,6 +161,7 @@ const getAllBlocks = () => {
 		accordionPanel,
 		archives,
 		audio,
+		breadcrumbs,
 		button,
 		buttons,
 		calendar,
@@ -217,7 +218,6 @@ const getAllBlocks = () => {
 		postContent,
 		postAuthor,
 		postAuthorName,
-		postBreadcrumbs,
 		postComment,
 		postCommentsCount,
 		postCommentsLink,
