+++ packages/block-library/src/style.scss
@@ -4,6 +4,7 @@
 @import "./archives/style.scss";
 @import "./avatar/style.scss";
 @import "./audio/style.scss";
+@import "./breadcrumbs/style.scss";
 @import "./button/style.scss";
 @import "./buttons/style.scss";
 @import "./calendar/style.scss";
@@ -38,7 +39,6 @@
 @import "./paragraph/style.scss";
 @import "./post-author/style.scss";
 @import "./post-author-biography/style.scss";
-@import "./post-breadcrumbs/style.scss";
 @import "./post-comments-form/style.scss";
 @import "./post-comments-count/style.scss";
 @import "./post-content/style.scss";
