+++ test/integration/fixtures/blocks/core__breadcrumbs.html
@@ -1,0 +1,1 @@
+<!-- wp:breadcrumbs /-->
