+++ test/integration/fixtures/blocks/core__breadcrumbs.serialized.html
@@ -1,0 +1,1 @@
+<!-- wp:breadcrumbs /-->
