--- test/integration/fixtures/blocks/core__post-breadcrumbs.html
@@ -1,1 +1,0 @@
-<!-- wp:post-breadcrumbs /-->
