--- test/integration/fixtures/blocks/core__post-breadcrumbs.serialized.html
@@ -1,1 +1,0 @@
-<!-- wp:post-breadcrumbs /-->
